+++ crates/clitest-lib/src/parser/v0/mod.rs
@@ -2,3 +2,7 @@ mod parse;
 mod segment;
 
 pub use parse::parse_script;
+
+const REGEX_MULTILINE: &str = "???";
+const ESCAPED_MULTILINE: &str = "!!!";
+const LITERAL_MULTILINE: &str = r#"""""#;