+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-08-11
+
+### Added
+- `not` patterns: Added support for negative lookahead patterns
+
 ## [0.2.0] - 2025-08-04
 
 ### Added